Refactorings needed for injection support
Different SCORE users might have specific needs for supporting
proprietary or copyrighted hashing functions. Therefore we need
to provide the mechanism for clients injecting their extensions
to the hash lib.

This commit prepares the open-source code for the injection of
a copyrighted hash function by one of the clients.

Notice that the aim is to achieve a design that will be
sufficiently generic that will allow for extensions without any
change on the open-source side.
