+++ score/hash/code/core/factory/impl/safe_hash_calculator_factory_ieee.cpp
@@ -38,7 +38,7 @@ Result<std::unique_ptr<IHashCalculator>> SafeHashCalculatorFactory::CreateHashCa
         case HashAlgorithm::kSha256:
             result = static_cast<std::unique_ptr<IHashCalculator>>(std::make_unique<Sha256Digest>());
             break;
-        case HashAlgorithm::kCrc32Unused:
+        case HashAlgorithm::kCrc32Autosar:
         case HashAlgorithm::kSha1:
         case HashAlgorithm::kSha384:
         case HashAlgorithm::kSha512:
