+++ score/hash/code/crc/crc32_ieee.cpp
@@ -12,6 +12,7 @@
  ********************************************************************************/
 
 #include "score/hash/code/crc/crc32_ieee.h"
+#include "score/hash/code/crc/lookup_table.h"
 
 #include <cstdint>
 #include <numeric>
@@ -26,47 +27,7 @@ namespace
 
 constexpr std::uint_fast32_t kAllOnes{0xFFFFFFFFU};
 
-template <std::uint_fast32_t ReversePolynomial>
-class LookupTable final
-{
-  public:
-    constexpr LookupTable() : table_{}
-    {
-        for (std::uint_fast32_t table_index = 0U; table_index < kTableSize; ++table_index)
-        {
-            auto checksum = table_index;
-
-            for (auto round = 0U; round < 8U; ++round)
-            {
-                if (static_cast<bool>(checksum & 0x1U))
-                {
-                    checksum = (checksum >> 1U) ^ ReversePolynomial;
-                }
-                else
-                {
-                    checksum = (checksum >> 1U) ^ 0U;
-                }
-            }
-            // NOLINTNEXTLINE(cppcoreguidelines-pro-bounds-constant-array-index) can't use .at() in constexpr fn
-            table_[table_index] = checksum;
-        }
-    }
-    constexpr std::uint_fast32_t operator[](size_t i) const noexcept
-    {
-        // NOLINTNEXTLINE(cppcoreguidelines-pro-bounds-constant-array-index) can't use .at() in constexpr fn
-        return table_[i];
-    }
-
-  private:
-    static constexpr auto kTableSize = 256U;
-
-    /* Doesn't impose security issues when used C-style array */
-    /* Using C-style array for filling array constexpr function (this can't be done with an array or a vector)*/
-    // NOLINTNEXTLINE(modernize-avoid-c-arrays) Using C-style array for filling array constexpr function
-    std::uint_fast32_t table_[kTableSize];
-};
-
-constexpr LookupTable<0xEDB88320U> kIeeeCrc32LookupTable{};
+constexpr internal::LookupTable<0xEDB88320U> kIeeeCrc32LookupTable{};
 
 template <typename InputIterator>
 constexpr std::uint_fast32_t CalculateIeeeCrc32(std::uint_fast32_t start, InputIterator first, InputIterator last)
